\name{SpMeta-package}
\alias{SpMeta-package}
\alias{SpMeta}
\docType{package}
\title{
\packageTitle{SpMeta}
}
\description{
\packageDescription{SpMeta}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SpMeta}

Maintainer: \packageMaintainer{SpMeta}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SpMeta)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SpMeta}}
%\section{Documentation Index}{\packageIndices{SpMeta}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
